requirements.txt
acres==0.3.0
annotated-types==0.7.0
asgiref==3.8.1
blis==1.3.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
cymem==2.0.11
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
etelemetry==0.3.1
faiss-cpu==1.10.0
filelock==3.18.0
fitz==0.0.1.dev2
fsspec==2025.3.2
httplib2==0.22.0
huggingface-hub==0.30.2
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
isodate==0.6.1
Jinja2==3.1.6
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
looseversion==1.3.0
lxml==5.3.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pgvector==0.4.0
pillow==11.2.1
preshed==3.0.9
prov==2.0.1
psycopg2-binary==2.9.10
puremagic==1.28
pydantic==2.11.3
pydantic_core==2.33.1
pydot==3.0.4
Pygments==2.19.1
PyJWT==2.9.0
PyMuPDF==1.25.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rank-bm25==0.2.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
sympy==1.13.1
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
traits==7.0.2
transformers==4.51.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
